Entropics.jl

Entropics — Module

Entropics.jl

Entropics.jl is a Julia package for computing the maximum-entropy distribution bounded on a close interval with given median, mean, and variance values.

The package in still under development.

An article introducing this package and its interesting applications is also in preparation.

Examples

The most important function provided in Entropics.jl is maxendist, which computes a maximum-entropy distribution with given conditions. For example, we may find the maximum-entropy distribution on interval $[0,1]$ with median $0.3$ and mean $0.35$, and then examine its median, mean, variance, skewness, kurtosis, and entropy.

julia> using Entropics

julia> d = maxendist(0, 1; median=0.3, mean=0.35)
Entropics.MED110{Float64}(0.0, 1.0, 0.3, -2.134967865645346, 0.8140360297855087, 0.9598690412098198)

julia> median(d)
0.3

julia> mean(d)
0.35

julia> var(d)
0.06861280728114773

julia> skewness(d)
0.6318688292278568

julia> kurtosis(d)
2.405231674579595

julia> entropy(d)
-0.13971378252179312

Mathematical base

Entropics.binaryroot — Function

binaryroot(f::Function, a::Real, b::Real)

Solve the equation $f(x) = 0$ using binary search method.

Entropics.secantroot — Function

secantroot(f::Function, a::Real, b::Real)

Solve the equation $f(x) = 0$ using secant method from given initial values.

Entropics.integrate — Function

integrate(f, a, b; abstol=1e-12, reltol=1e-12)

Compute $\int_a^b f(x) \operatorname{d}x$ with Cuhre algorithm in Cuba.jl.

Entropics.xlog — Function

xlog(x::AbstractFloat)
xlog(x::Real)

Compute $x * \log(x)$, with a convention at $x=0$ (returning $0$).

Entropics.erfiinv — Function

erfiinv(y::AbstractFloat)
erfiinv(y::Real)

The inverse of erfi, finding $x$ satisfying $y = \operatorname{erfi}(x)$.

Entropics.cothminv — Function

cothminv(x::T)
cothminv(x::Real)

Compute $\coth(x) - 1/x$ in a numerically stable scheme.

Entropics.diffexp — Function

diffexp(b::T, a::T) where {T<:AbstractFloat}
diffexp(b::AbstractFloat, a::AbstractFloat)
diffexp(b::Real, a::Real)

Compute $\exp(b) - \exp(a)$ in a numerically stable scheme.

Entropics.diffxexp — Function

diffxexp(b::T, a::T) where {T<:AbstractFloat}
diffxexp(b::AbstractFloat, a::AbstractFloat)
diffxexp(b::Real, a::Real)

Compute $b \exp(b) - a \exp(a)$ in a numerically stable scheme.

Entropics.diffxxexp — Function

diffxxexp(b::T, a::T) where {T<:AbstractFloat}
diffxxexp(b::AbstractFloat, a::AbstractFloat)
diffxxexp(b::Real, a::Real)

Compute $b^2 \exp(b) - a^2 \exp(a)$ in a numerically stable scheme.

Entropics.diffxexpf — Function

diffxexpf(f::Function, b::T, a::T) where {T<:AbstractFloat}
diffxexpf(f::Function, b::AbstractFloat, a::AbstractFloat)
diffxexpf(f::Function, b::Real, a::Real)

Compute $b \exp(f(b)) - a \exp(f(a))$ in a numerically stable scheme.

Entropics.differf — Function

differf(b::T, a::T) where {T<:AbstractFloat}
differf(b::AbstractFloat, a::AbstractFloat)
differf(b::Real, a::Real)

Compute $\erf(b) - \erf(a)$ in a numerically stable scheme.

Entropics.differfi — Function

differfi(b::T, a::T) where {T<:AbstractFloat}
differfi(b::AbstractFloat, a::AbstractFloat)
differfi(b::Real, a::Real)

Compute $\erfi(b) - \erfi(a)$ in a numerically stable scheme.

Entropics.intexp — Function

intexp(A::T, B::T, C::T, a::T, b::T) where {T<:AbstractFloat}
intexp(A::AbstractFloat, B::AbstractFloat, C::AbstractFloat, 
	a::AbstractFloat, b::AbstractFloat)
intexp(A::Real, B::Real, C::Real, a::Real, b::Real)

Compute $\int_a^b \exp(A*x^2 + B*x + C) \operatorname{d}x$.

Entropics.intxexp — Function

intxexp(A::T, B::T, C::T, a::T, b::T) where {T<:AbstractFloat}
intxexp(A::AbstractFloat, B::AbstractFloat, C::AbstractFloat, 
	a::AbstractFloat, b::AbstractFloat)
intxexp(A::Real, B::Real, C::Real, a::Real, b::Real)

Compute $\int_a^b x * \exp(A*x^2 + B*x + C) \operatorname{d}x$.

Entropics.intxxexp — Function

intxxexp(A::T, B::T, C::T, a::T, b::T) where {T<:AbstractFloat}
intxxexp(A::AbstractFloat, B::AbstractFloat, C::AbstractFloat, 
	a::AbstractFloat, b::AbstractFloat)
intxxexp(A::Real, B::Real, C::Real, a::Real, b::Real)

Compute $\int_a^b x^2 * \exp(A*x^2 + B*x + C) \operatorname{d}x$.

Entropics.invintexp — Function

invintexp(A::T, B::T, C::T, a::T, p::T) where {T<:AbstractFloat}
invintexp(A::AbstractFloat, B::AbstractFloat, C::AbstractFloat, 
	a::AbstractFloat, p::AbstractFloat)
invintexp(A::Real, B::Real, C::Real, a::Real, p::Real)

The inverse of intexp(A, B, C, a, b) with respect to b, which finds $b$ satisfying $p = \int_a^b \exp(A*x^2 + B*x + C) \operatorname{d}x$.

Entropics.intqexp — Function

intqexp(A::T, B::T, C::T, a::T, b::T) where {T<:AbstractFloat}
intqexp(A::AbstractFloat, B::AbstractFloat, C::AbstractFloat, 
	a::AbstractFloat, b::AbstractFloat)
intqexp(A::Real, B::Real, C::Real, a::Real, b::Real)

Compute $- \int_a^b f(x) * \exp(f(x)) \operatorname{d}x$, where $f(x) = A*x^2 + B*x + C$.

Types

Entropics.PDF — Type

PDF{D<:Distribution}

Container type for a probability density function.

Entropics.pdf — Function

pdf(d::Distribution)

Get the probability density function of a distribution as a Julia function.

Entropics.pab — Function

pab(p::Function, x::Real, a::Real, b::Real)

Return

zero, if x < a,

p(x), if a <= x <= b,

zero, if x > b.

Entropics.pamb — Function

pamb(pa::Function, pb::Function, x::Real, a::Real, m::Real, b::Real)

Return

zero, if x < a,

pa(x), if a <= x <= m,

pb(x), if m < x <= b,

zero, if x > b.

Entropics.CDF — Type

CDF{D<:Distribution}

Container type for a cumulative distribution function.

Entropics.Pab — Function

Pab(P::Function, x::Real, a::Real, b::Real)

Return

zero, if x < a,

P(x), if a <= x <= b,

one, if x > b.

Entropics.Pamb — Function

Pamb(Pa::Function, Pb::Function, x::Real, a::Real, m::Real, b::Real)

Return

zero, if x < a,

Pa(x), if a <= x <= m,

Pb(x) + 1/2, if m < x <= b,

one, if x > b.

Entropics.quantile — Function

quantile(d::Distribution, p::Real, a::Real, b::Real)
quantile(d::Distribution, p::Real)
quantile(d::Distribution, pp::Vector{<:Real})
quantile(d::Distribution)

Compute the quantile value in a distribution. The last syntax creates a Julia function similar to those from pdf and cdf that maps p to q(d, p).

Entropics.q01 — Function

q01(q::Function, p::Real)

Return q(p) if 0 <= p <= 1, or an error otherwise.

Entropics.support — Function

support(d::Distribution)

Compute the support (a, b) of a distribution. For unbounded distributions, the support interval expands from negative infinity to positive infinity.

Entropics.median — Function

median(d::Distribution)

Compute the median of a distribution.

Entropics.mean — Function

mean(d::Distribution)

Compute the mean of a distribution.

Entropics.moment2 — Function

moment2(d::Distribution)

Compute the second moment (e.g. $\mathcal{E}[X^2]$ for a random variable $X$) of a distribution.

Entropics.var — Function

var(d::Distribution; u=mean(d))

Compute the variance of a distribution. The mean can be provided if known.

Entropics.std — Function

std(d::Distribution; u=mean(d))

Compute the standard deviation of a distribution. The mean can be provided if known.

Entropics.entropy — Function

entropy(d::Bounded; u=mean(d))

Compute the differential entropy of a bounded distribution.

Entropics.skewness — Function

skewness(d::Bounded; u=mean(d), s=std(d))

Compute the skewness of a bounded distribution.

Entropics.kurtosis — Function

kurtosis(d::Bounded; u=mean(d), s=std(d))

Compute the kurtosis of a bounded distribution.

Entropics.bound — Function

bound(d::Distribution, a::Real, b::Real)
bound(d::Bounded)
bound(d::Smooth{T,<:Sample{T}}) where {T<:AbstractFloat}

Bound or truncate a distribution by interval $[a, b]$, and then return the new bounded distribution. Especially, bounding a bounded distribution without giving a and b returns the bounded distribution itself; when bounding a smoothed sample (see smooth) without giving a and b, the interval is set to the extrema of the original sample.

Entropics.smooth — Function

smooth(d::Distribution, h::Real)
smooth(d::Smooth)
smooth(d::Sample)
smooth(v::Vector{<:Real})

Smooth a distribution with a Gaussian kernal with bandwidth h. Especially, smoothing a smoothed distribution without giving h returns the smoothed distribution itself; when smoothing a sample or a vector of reals, the bandwidth is calculated using Silverman (1986)'s rule of thumb (see helrot for more details).

Maximum-entropy distributions

Entropics.urange — Function

urange(a::Real, b::Real, m::Real, v::Real)

Return the possible range of mean of a distribution bounded on $[a, b]$ with median m and variance v as an interval $(ul, ur)$.

Entropics.maxendist — Function

maxendist(a::Real, b::Real; 
	median::Union{Real,Nothing}=nothing, 
	mean::Union{Real,Nothing}=nothing, 
	std::Union{Real,Nothing}=nothing, var::Union{Real,Nothing}=nothing

Compute the maximum-entropy distribution bounded on $[a, b]$ with given median, mean, standard deviation, and/or variance.

Entropics.xab — Function

xab(a::Real, b::Real)

Check if $[a, b]$ is a well-defined real interval with positive length.

Entropics.xabm — Function

xabm(a::Real, b::Real, m::Real)

Check if median m is possible for a continuous distribution bounded on $[a, b]$. See Theorem 100 in the article.

Entropics.xabu — Function

xabu(a::Real, b::Real, u::Real)

Check if mean u is possible for a continuous distribution bounded on $[a, b]$. See Theorem 010 in the article.

Entropics.xs — Function

xs(s::Real)

Check if standard deviation s is possible for a continuous distribution.

Entropics.xv — Function

xv(v::Real)

Check if variance v is possible for a continuous distribution.

Entropics.xsv — Function

xsv(s::Real, v::Real)

Check if standard deviation s and variance v is consistent for a distribution. If so, the variance is returned.

Entropics.xabuv — Function

xabuv(a::Real, b::Real, u::Union{Real, Nothing}, v::Union{Real, Nothing})

Check if mean u and variance v is simultaneously possible for a continuous distribution bounded on $[a, b]$. See Theorem 001 (Popoviciu(1935)'s inequality) and Theorem 011 (Bhatia & Davis (2000)'s inequality) in the article.

Entropics.xabmu — Function

xabmu(a::Real, b::Real, m::Union{Real, Nothing}, u::Union{Real, Nothing})

Check if median m and mean u is simultaneously possible for a continuous distribution bounded on $[a, b]$. See Theorem 110 in the article.

Entropics.xabmuv — Function

xabmuv(a::Real, b::Real, m::Union{Real, Nothing}, 
	u::Union{Real, Nothing}, v::Union{Real, Nothing})

Check if median m, mean u, and variance v is simultaneously possible for a continuous distribution bounded on $[a, b]$. See Theorem 111a and Theorem 111b in the article.

Sampling

Entropics.sample — Function

sample(v::Vector{<:Real})

Create a sample as a distribution.

Entropics.helrot — Function

helrot(d::Sample{T}; useiqr::Bool=true) where {T<:AbstractFloat}

Compute the optimal bandwidth using Silverman (1986)'s rule of thumb for smoothing with a Gaussian kernal. If useiqr is true, then the version using interquartile range (IQR) is applied.